ai-analysis-service/requirements.runtime.txt
confluent-kafka>=2.0.0
avro>=1.11.0
pydantic>=2.0.0
fastavro>=1.7.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
prometheus-client>=0.20.0
redis>=5.0.0
websockets>=12.0
msgpack>=1.0.8
